requirements.txt
streamlit>=1.37.0
PyMuPDF>=1.24.0
opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=1.26.0
